cleaned up and finished auth
