+++ app/android/src/main/java/com/example/tutorial/android/ui/main/HomeActivity.kt
@@ -1,6 +1,7 @@
 package com.example.tutorial.android.ui.main
 
 import android.os.Bundle
+import android.util.Log
 import com.google.android.material.bottomnavigation.BottomNavigationView
 import androidx.appcompat.app.AppCompatActivity
 import androidx.navigation.findNavController
@@ -9,7 +10,7 @@ import androidx.navigation.ui.setupActionBarWithNavController
 import androidx.navigation.ui.setupWithNavController
 import com.example.tutorial.android.R
 
-class HomeActivity : AppCompatActivity() {
+class HomeActivity() : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -17,6 +18,10 @@ class HomeActivity : AppCompatActivity() {
             this.supportActionBar!!.hide()
         } catch (e: NullPointerException) {
             println(e)
+        }
+        val bundle = intent.extras
+        if (bundle != null) {
+            Log.d("CHOSEN TRAINING", bundle.get("training") as String)
         }
         setContentView(R.layout.activity_home)
         val navView: BottomNavigationView = findViewById(R.id.nav_view)
